pyproject.toml
[project]
name = "beeb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "jinja2-fragments>=1.9.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "ruff>=0.11.13",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]
